# docs/research/mvp/final/implementation/api/openapi.yaml
openapi: "3.1.0"
info:
  title: HelixTerminator API
  version: "1.0.0"
  description: |
    HelixTerminator — Next-Generation Enterprise SSH Client Platform.
    Dual-module product family: Module A (Secure Terminal Platform) + Module B (Zero-Trust Connection Broker).
  contact:
    name: HelixTerminator Engineering
    email: api@helixterminator.io
    url: https://helixterminator.io
  license:
    name: Proprietary
    identifier: Proprietary
servers:
  - url: https://api.helixterminator.io/api/v1
    description: Production
  - url: https://staging-api.helixterminator.io/api/v1
    description: Staging
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      summary: Register new user
      operationId: authRegister
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimit' }
  /auth/login:
    post:
      summary: Authenticate user
      operationId: authLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated or MFA required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimit' }
  /auth/logout:
    post:
      summary: Logout
      operationId: authLogout
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                all_sessions: { type: boolean, default: false }
      responses:
        '204': { description: Logged out }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      summary: Refresh access token
      operationId: authRefresh
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: New tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /auth/sessions/{sessionId}:
    delete:
      summary: Terminate session
      operationId: authTerminateSession
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Session terminated }

  # User Service
  /users/me:
    get:
      summary: Get current user
      operationId: userGetMe
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      summary: Update user profile
      operationId: userUpdateMe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      summary: Delete account
      operationId: userDeleteMe
      responses:
        '204': { description: Account scheduled for deletion }
  /users/me/preferences:
    get:
      summary: Get preferences
      operationId: userGetPreferences
      responses:
        '200': { description: Preferences object }
    put:
      summary: Update preferences
      operationId: userUpdatePreferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200': { description: Updated preferences }
...
  /vaults:
    get:
      summary: List vaults
      operationId: vaultList
      parameters:
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 25 } }
      responses:
        '200':
          description: Vault list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    post:
      summary: Create vault
      operationId: vaultCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VaultCreateRequest'
      responses:
        '201':
          description: Vault created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vault'
  /vaults/{vaultId}:
    get:
      summary: Get vault
      operationId: vaultGet
      parameters:
        - { name: vaultId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Vault
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vault'
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Update vault
      operationId: vaultUpdate
      parameters:
        - { name: vaultId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VaultUpdateRequest'
      responses:
        '200':
          description: Updated vault
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vault'
    delete:
      summary: Delete vault
      operationId: vaultDelete
      parameters:
        - { name: vaultId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Vault deleted }
  /vaults/{vaultId}/items:
    get:
      summary: List vault items
      operationId: vaultListItems
      parameters:
        - { name: vaultId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 25 } }
      responses:
        '200':
          description: Item list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    post:
      summary: Create vault item
      operationId: vaultCreateItem
      parameters:
        - { name: vaultId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VaultItemCreateRequest'
      responses:
        '201':
          description: Item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaultItem'
...
  /hosts:
    get:
      summary: List hosts
      operationId: hostList
      parameters:
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 25 } }
        - { name: status, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
      responses:
        '200':
          description: Host list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    post:
      summary: Create host
      operationId: hostCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostCreateRequest'
      responses:
        '201':
          description: Host created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Host'
  /hosts/{hostId}:
    get:
      summary: Get host
      operationId: hostGet
      parameters:
        - { name: hostId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Host
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Host'
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Update host
      operationId: hostUpdate
      parameters:
        - { name: hostId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostUpdateRequest'
      responses:
        '200':
          description: Updated host
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Host'
    delete:
      summary: Delete host
      operationId: hostDelete
      parameters:
        - { name: hostId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Host deleted }
  /hosts/{hostId}/connect:
    post:
      summary: Connect to host
      operationId: hostConnect
      parameters:
        - { name: hostId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                key_id: { type: string, format: uuid }
                auth_method: { type: string }
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
...
  /sessions:
    get:
      summary: List sessions
      operationId: sessionList
      parameters:
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 25 } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        '200':
          description: Session list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    post:
      summary: Create session
      operationId: sessionCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionCreateRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
  /sessions/{sessionId}:
    get:
      summary: Get session
      operationId: sessionGet
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
    delete:
      summary: Terminate session
      operationId: sessionTerminate
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Session terminated }
...
  /snippets:
    get:
      summary: List snippets
      operationId: snippetList
      parameters:
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 25 } }
        - { name: q, in: query, schema: { type: string } }
      responses:
        '200':
          description: Snippet list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    post:
      summary: Create snippet
      operationId: snippetCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SnippetCreateRequest'
      responses:
        '201':
          description: Snippet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snippet'
...
  /workspaces/{workspaceId}/sessions:
    get:
      summary: Sessions in workspace
      operationId: workspaceSessions
      parameters:
        - { name: workspaceId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Session list }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: EdDSA (Ed25519) signed JWT
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, display_name, accept_terms]
      properties:
        email: { type: string, format: email, maxLength: 255 }
        password: { type: string, minLength: 12 }
        display_name: { type: string, minLength: 2, maxLength: 100 }
        invite_code: { type: string }
        accept_terms: { type: boolean }
        locale: { type: string, default: "en-US" }
        timezone: { type: string, default: "UTC" }
...
  responses:
...
    Conflict:
      description: Conflict
      content:
        application/problem+json:
          schema:
            type: object
            properties:
              type: { type: string }
              title: { type: string }
              status: { type: integer }
              detail: { type: string }
...